+++ chunkflow/chunk/base.py
@@ -144,6 +144,7 @@ class Chunk(NDArrayOperatorsMixin):
 
         if dtype:
             arr = arr.astype(dtype)
+        print(f'read tif chunk with size of {arr.shape}, voxel offset: {voxel_offset}, voxel size: {voxel_size}')
         return cls(arr, voxel_offset=voxel_offset, voxel_size=voxel_size)
     
     def to_tif(self, file_name: str=None):
@@ -231,7 +232,9 @@ ends with {cutout_stop}, size is {cutout_size}, voxel size is {voxel_size}.""")
 
     def to_h5(self, file_name: str, with_offset: bool=True, 
                 chunk_size: tuple=(64,64,64),
-                with_unique: bool= True, compression="gzip"):
+                with_unique: bool= True, 
+                compression="gzip",
+                voxel_size: tuple = None):
         """
         :param file_name: output file name. If it is not end with h5, the coordinate will be appended to the file name.
         :param with_offset: save the voxel offset or not
@@ -252,8 +255,11 @@ ends with {cutout_stop}, size is {cutout_size}, voxel size is {voxel_size}.""")
 
         with h5py.File(file_name, 'w') as f:
             f.create_dataset('/main', data=self.array, chunks=chunk_size, compression=compression)
-            if self.voxel_size:
-                f.create_dataset('/voxel_size', data=self.voxel_size)
+            if voxel_size is None and self.voxel_size is not None:
+                voxel_size = self.voxel_size
+            if voxel_size is not None:
+                f.create_dataset('/voxel_size', data=voxel_size)
+
             if with_offset and self.voxel_offset is not None:
                 f.create_dataset('/voxel_offset', data=self.voxel_offset)
 
@@ -262,6 +268,7 @@ ends with {cutout_stop}, size is {cutout_size}, voxel size is {voxel_size}.""")
                 if unique[0]:
                     unique = unique[1:]
                 f.create_dataset('/unique_nonzeros', data = unique)
+        return file_name
 
     def __array__(self):
         return self.array
@@ -338,7 +345,7 @@ ends with {cutout_stop}, size is {cutout_size}, voxel size is {voxel_size}.""")
         if self.voxel_size is not None or self.voxel_size != (1, 1, 1):
             props['voxel_size'] = self.voxel_size
         return props 
-
+    
     @property
     def slices(self) -> tuple:
         """
@@ -426,6 +433,7 @@ ends with {cutout_stop}, size is {cutout_size}, voxel size is {voxel_size}.""")
         else:
             voxel_size = self.voxel_size
         return Chunk(new_array, voxel_offset=voxel_offset, voxel_size=voxel_size)
+
 
     def fill(self, x):
         self.array.fill(x)