+++ CHANGELOG.md
@@ -6,13 +6,14 @@ ChangeLog history
 - neuroglancer operator input changed from `chunk_names` to `inputs` to be consistent with other layers.
 - move agglomerate operator to a plugin to remove dependency of waterz. Waterz is not easy to install in MAC os and is not currently supported in python 3.9. After this change, we should be able to support python 3.9 now.
 - replaced evaluate code from waterz to gala. The evaluate code is copied from gala following the license permission. This was also got agreement from the package author [Juan](https://github.com/janelia-flyem/gala/issues/99#issuecomment-823767079)
-
+- change default connectivity from 26 to 6 in connected-components operator
 ## Deprecations 
 
 ## Features
 - added synapse annotation visualization.
 - The code of neuroglancer operator is splitted to functions and becomes more modular. The code is more readable and maintainable.
 ## Bug Fixes 
+- fix chunk transpose with None voxel size and voxel offset
 
 ## Improved Documentation